requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
pinecone-client==5.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
simple-salesforce==1.12.6
python-dotenv==1.0.1
tiktoken==0.7.0
